+++ extension/lib/factfinder/client/errors/FactFinderInvalidResponseError.js
@@ -1,8 +1,12 @@
 'use strict'
 
 class FactFinderInvalidResponseError extends Error {
-  constructor () {
+  /**
+   * @param {Object} response
+   */
+  constructor (response) {
     super()
+    this.response = response
     this.message = `Empty or invalid response from FACT-Finder`
   }
 }